feat: Implement prompt to display raw data
A function raw_data(df) is created which asks user if they want to see 5 lines of raw data,
and if the user wants to see data, it displays 5 lines of data and asks again if they want to
see the next 5 lines. This continues until user says no.

raw_data(df):
	Displays first 5 lines of Raw Data.